+++ rio_tiler/mosaic/reader.py
@@ -3,11 +3,17 @@
 from inspect import isclass
 from typing import Any, Callable, List, Optional, Sequence, Tuple, Type, Union, cast
 
+import numpy
 from rasterio.crs import CRS
 
 from rio_tiler.constants import MAX_THREADS
-from rio_tiler.errors import EmptyMosaicError, InvalidMosaicMethod, TileOutsideBounds
-from rio_tiler.models import ImageData
+from rio_tiler.errors import (
+    EmptyMosaicError,
+    InvalidMosaicMethod,
+    PointOutsideBounds,
+    TileOutsideBounds,
+)
+from rio_tiler.models import ImageData, PointData
 from rio_tiler.mosaic.methods.base import MosaicMethodBase
 from rio_tiler.mosaic.methods.defaults import FirstMethod
 from rio_tiler.tasks import create_tasks, filter_tasks
@@ -118,6 +124,104 @@ def mosaic_reader(
             assets=assets_used,
             crs=crs,
             bounds=bounds,
+            band_names=band_names,
+        ),
+        assets_used,
+    )
+
+
+def mosaic_point_reader(
+    mosaic_assets: Sequence,
+    reader: Callable[..., PointData],
+    *args: Any,
+    pixel_selection: Union[Type[MosaicMethodBase], MosaicMethodBase] = FirstMethod,
+    chunk_size: Optional[int] = None,
+    threads: int = MAX_THREADS,
+    allowed_exceptions: Tuple = (PointOutsideBounds,),
+    **kwargs,
+) -> Tuple[PointData, List]:
+    """Merge multiple assets.
+
+    Args:
+
+        mosaic_assets (sequence): List of assets.
+        reader (callable): Reader function. The function MUST take `(asset, *args, **kwargs)` as arguments, and MUST return a PointData object.
+        args (Any): Argument to forward to the reader function.
+        pixel_selection (MosaicMethod, optional): Instance of MosaicMethodBase class. Defaults to `rio_tiler.mosaic.methods.defaults.FirstMethod`.
+        chunk_size (int, optional): Control the number of asset to process per loop.
+        threads (int, optional): Number of threads to use. If <= 1, runs single threaded without an event loop. By default reads from the MAX_THREADS environment variable, and if not found defaults to multiprocessing.cpu_count() * 5.
+        allowed_exceptions (tuple, optional): List of exceptions which will be ignored. Note: `TileOutsideBounds` is likely to be raised and should be included in the allowed_exceptions. Defaults to `(TileOutsideBounds, )`.
+        kwargs (optional): Reader callable's keywords options.
+
+    Returns:
+        tuple: ImageData and assets (list).
+
+    Examples:
+        >>> def reader(asset: str, *args, **kwargs) -> PointData:
+                with Reader(asset) as src:
+                    return src.point(*args, **kwargs)
+
+            pt = mosaic_point_reader(["cog.tif", "cog2.tif"], reader, 0, 0)
+
+    """
+    if isclass(pixel_selection):
+        pixel_selection = cast(Type[MosaicMethodBase], pixel_selection)
+
+        if issubclass(pixel_selection, MosaicMethodBase):
+            pixel_selection = pixel_selection()
+
+    if not isinstance(pixel_selection, MosaicMethodBase):
+        raise InvalidMosaicMethod(
+            "Mosaic filling algorithm should be an instance of "
+            "'rio_tiler.mosaic.methods.base.MosaicMethodBase'"
+        )
+
+    if not chunk_size:
+        chunk_size = threads if threads > 1 else len(mosaic_assets)
+
+    assets_used: List = []
+    crs: Optional[CRS]
+    coordinates: Optional[Tuple[float, float]]
+    band_names: List[str]
+
+    for chunks in _chunks(mosaic_assets, chunk_size):
+        tasks = create_tasks(reader, chunks, threads, *args, **kwargs)
+        for pt, asset in filter_tasks(
+            tasks,
+            allowed_exceptions=allowed_exceptions,
+        ):
+            crs = pt.crs
+            coordinates = pt.coordinates
+            band_names = pt.band_names
+
+            assets_used.append(asset)
+            pixel_selection.feed(pt.as_masked())
+
+            if pixel_selection.is_done:
+                data, mask = pixel_selection.data
+                return (
+                    PointData(
+                        data,
+                        numpy.array([mask], dtype="uint8"),
+                        assets=assets_used,
+                        crs=crs,
+                        coordinates=coordinates,
+                        band_names=band_names,
+                    ),
+                    assets_used,
+                )
+
+    data, mask = pixel_selection.data
+    if data is None:
+        raise EmptyMosaicError("Method returned an empty array")
+
+    return (
+        PointData(
+            data,
+            numpy.array([mask], dtype="uint8"),
+            assets=assets_used,
+            crs=crs,
+            coordinates=coordinates,
             band_names=band_names,
         ),
         assets_used,